+++ android/src/main/kotlin/com/audienzz/audienzz_sdk_flutter/AudienzzSdkFlutterPlugin.kt
@@ -69,7 +69,7 @@ class AudienzzSdkFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler
                 val bannerAd = BannerAd(
                     call.argument<String>("adUnitId")!!,
                     call.argument<String>("auConfigId")!!,
-                    call.argument<AdSize>("adSize")!!,
+                    call.argument<List<AdSize>>("adSizes")!!,
                     call.argument<Boolean>("isAdaptiveSize")!!,
                     call.argument<Int?>("refreshTimeInterval"),
                     call.argument<AdFormat>("adFormat")!!,
@@ -81,6 +81,7 @@ class AudienzzSdkFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler
                     call.argument<VideoDuration>("videoDuration")!!,
                     call.argument<String?>("pbAdSlot"),
                     call.argument<String?>("gpId"),
+                    call.argument<String?>("impOrtbConfig"),
                     adInstanceManager?.createBannerAdListener(adId),
                     context,
                 )
@@ -104,6 +105,7 @@ class AudienzzSdkFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler
                     call.argument<VideoDuration>("videoDuration")!!,
                     call.argument<String?>("pbAdSlot"),
                     call.argument<String?>("gpId"),
+                    call.argument<String?>("impOrtbConfig"),
                     context,
                     adInstanceManager!!.createRewardedAdLoadedListener(adId),
                     adInstanceManager!!.createRewardedAdUserEarnedRewardListener(adId),
@@ -131,6 +133,8 @@ class AudienzzSdkFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler
                     call.argument<VideoDuration>("videoDuration")!!,
                     call.argument<String?>("pbAdSlot"),
                     call.argument<String?>("gpId"),
+                    call.argument<List<AdSize>>("adSizes")!!,
+                    call.argument<String?>("impOrtbConfig"),
                     context,
                     adInstanceManager!!.createInterstitialAdLoadedListener(adId),
                     adInstanceManager!!.createOverlayAdFullscreenContentListener(adId),
@@ -151,6 +155,17 @@ class AudienzzSdkFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler
                 }
 
                 result.success(null)
+            }
+
+            "getPlatformAdSize" -> {
+                val adId = call.argument<Int>("adId")!!
+                val ad = adInstanceManager?.adFor(adId)
+
+                if(ad is BannerAd) {
+                    result.success(ad.getPlatformAdSize())
+                } else {
+                    result.success(null)
+                }
             }