+++ lib/src/entities/exceptions/ad_size_required_exception.dart
@@ -1,0 +1,3 @@
+final class AdSizeRequiredException implements Exception {
+  const AdSizeRequiredException();
+}